requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=0.19.0
pydantic>=1.8.0
numpy>=1.21.0
textblob>=0.15.3
vaderSentiment>=3.3.2
pytest==7.4.3
httpx==0.25.2
python-multipart==0.0.6
